refactor: introduce JenkinsFunSpec and JenkinsSpec base classes for Kotest and Spock
Each framework now has its own wrapper that encapsulates JenkinsRule lifecycle
(before/after) and the recipe() override in one place. Test files extend the base
class and access rule directly, with no per-file boilerplate. KotestJenkinsRule.kt
deleted.
